+++ src/main/java/net/silentchaos512/gear/item/gear/GearArrowItem.java
@@ -99,7 +99,7 @@ public class GearArrowItem extends ArrowItem implements GearItem {
     @Override
     public ItemAttributeModifiers getDefaultAttributeModifiers(ItemStack stack) {
         var builder = ItemAttributeModifiers.builder();
-        GearHelper.addAttributeModifiers(stack, builder);
+        GearHelper.addAttributeModifiers(stack, builder, false);
         return builder.build();
     }