+++ src/main/java/net/silentchaos512/gear/item/gear/GearShearsItem.java
@@ -83,7 +83,7 @@ public class GearShearsItem extends ShearsItem implements GearTool {
     @Override
     public ItemAttributeModifiers getDefaultAttributeModifiers(ItemStack stack) {
         var builder = ItemAttributeModifiers.builder();
-        GearHelper.addAttributeModifiers(stack, builder);
+        GearHelper.addAttributeModifiers(stack, builder, false);
         return builder.build();
     }