+++ src/main/java/net/silentchaos512/gear/item/gear/GearShieldItem.java
@@ -73,7 +73,7 @@ public class GearShieldItem extends ShieldItem implements GearItem {
     @Override
     public ItemAttributeModifiers getDefaultAttributeModifiers(ItemStack stack) {
         var builder = ItemAttributeModifiers.builder();
-        GearHelper.addAttributeModifiers(stack, builder);
+        GearHelper.addAttributeModifiers(stack, builder, false);
         return builder.build();
     }